scrap: rename coder.rs to codec.rs
