+++ libs/scrap/src/common/codec.rs
