+++ src/server/connection.rs
@@ -365,7 +365,7 @@ impl Connection {
         video_service::notify_video_frame_feched(id, None);
         super::video_service::update_test_latency(id, 0);
         super::video_service::update_image_quality(id, None);
-        scrap::coder::Encoder::update_video_encoder(id, None);
+        scrap::codec::Encoder::update_video_encoder(id, None);
         if let Err(err) = conn.try_port_forward_loop(&mut rx_from_cm).await {
             conn.on_close(&err.to_string(), false);
         }
@@ -1190,9 +1190,9 @@ impl Connection {
 
         // TODO: add option
         if let Some(q) = o.video_codec_state.clone().take() {
-            scrap::coder::Encoder::update_video_encoder(self.inner.id(), Some(q));
+            scrap::codec::Encoder::update_video_encoder(self.inner.id(), Some(q));
         } else {
-            scrap::coder::Encoder::update_video_encoder(
+            scrap::codec::Encoder::update_video_encoder(
                 self.inner.id(),
                 Some(VideoCodecState {
                     H264: false,
